pyproject.toml
[project]
name = "meu-tcc"
version = "0.1.0"
description = ""
authors = [
  {name = "Ermeson Alves", email = "ermeson3alves@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentationsx>=2.0.13",
    "ipywidgets>=8.1.8",
    "lightning>=2.6.0",
    "matplotlib>=3.10.8",
    "nibabel>=5.3.3",
    "numpy>=2.3.5",
    "opencv-python-headless>=4.11.0.86",
    "pynrrd>=1.1.3",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "segmentation-models-pytorch>=0.5.0",
    "simpleitk>=2.5.3",
    "tensorboard>=2.20.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.5.0",
]
